+++ apps/ui/src/components/Trace/span.tsx
@@ -3,7 +3,7 @@ import { UnixNanoTimeStamp } from "@graphql-debugger/time";
 import { Modal } from "../../context/modal";
 import { OpenModal } from "../modal/open";
 import { ModalWindow } from "../modal/window";
-import { JsonViewer } from "./editor/json/JsonViewer";
+import { JsonViewer } from "./editor/json/viewer";
 import { RenderTree } from "./utils";
 
 export function Span({
