+++ apps/ui/src/components/utils/highlighted-text.tsx
