+++ apps/ui/src/pages/Dashboard.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 
 import { InfoLogo } from "../components/info/info-logo";
-import { Page } from "../components/utils/Page";
+import { Page } from "../components/utils/page";
 import { SchemasContext } from "../context/schemas";
 
 export function Dashboard() {
